Move everything to dev.drewhamilton.rxpreferences package
Move everything to dev.drewhamilton.rxpreferences package
