+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <manifest
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="drewhamilton.rxpreferences3.example">
+    package="dev.drewhamilton.rxpreferences.example">
 
   <application
       android:label="@string/app_name"
