+++ app/src/main/java/dev/drewhamilton/rxpreferences/example/ApplicationModule.kt
@@ -1,0 +1,12 @@
+package dev.drewhamilton.rxpreferences.example
+
+import android.content.Context
+import dagger.Module
+import dagger.Provides
+
+@Module
+object ApplicationModule {
+
+    @Provides
+    @JvmStatic fun applicationContext(application: ExampleApplication): Context = application.applicationContext
+}
