+++ app/src/main/java/dev/drewhamilton/rxpreferences/example/MainActivity.kt
@@ -1,0 +1,12 @@
+package dev.drewhamilton.rxpreferences.example
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.main)
+    }
+}
