+++ app/src/main/java/dev/drewhamilton/rxpreferences/example/edit/EditingViewModel.kt
@@ -1,0 +1,15 @@
+package dev.drewhamilton.rxpreferences.example.edit
+
+import androidx.lifecycle.ViewModel
+import javax.inject.Inject
+
+class EditingViewModel @Inject constructor(private val exampleRepository: MutableExampleRepository) : ViewModel() {
+
+    fun getExampleStringOnce() = exampleRepository.getExampleStringOnce()
+
+    fun getExampleIntegerOnce() = exampleRepository.getExampleIntOnce()
+
+    fun setExampleValues(string: String, int: Int) = exampleRepository.changeExampleValues(string, int)
+
+    fun removeExampleValues() = exampleRepository.removeExampleValues()
+}
