--- app/src/main/java/drewhamilton/rxpreferences3/example/ApplicationComponent.kt
@@ -1,26 +1,0 @@
-package drewhamilton.rxpreferences3.example
-
-import dagger.BindsInstance
-import dagger.Component
-import drewhamilton.rxpreferences3.example.edit.EditingFragment
-import drewhamilton.rxpreferences3.example.observe.ObservationFragment
-
-@Component(modules = [
-    ApplicationModule::class,
-    PersistenceModule::class
-])
-interface ApplicationComponent {
-
-    fun inject(editingFragment: EditingFragment)
-
-    fun inject(observationFragment: ObservationFragment)
-
-    @Component.Factory
-    interface Factory {
-        fun create(@BindsInstance application: ExampleApplication): ApplicationComponent
-    }
-
-    companion object {
-        fun create(application: ExampleApplication) = DaggerApplicationComponent.factory().create(application)
-    }
-}
