--- app/src/main/java/drewhamilton/rxpreferences3/example/MainActivity.kt
@@ -1,12 +1,0 @@
-package drewhamilton.rxpreferences3.example
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.main)
-    }
-}
