--- app/src/main/java/drewhamilton/rxpreferences3/example/observe/ObservationViewModel.kt
@@ -1,11 +1,0 @@
-package drewhamilton.rxpreferences3.example.observe
-
-import androidx.lifecycle.ViewModel
-import javax.inject.Inject
-
-class ObservationViewModel @Inject constructor(private val exampleRepository: ExampleRepository) : ViewModel() {
-
-    fun getExampleStringStream() = exampleRepository.getExampleStringStream()
-
-    fun getExampleIntStream() = exampleRepository.getExampleIntStream()
-}
