+++ dagger/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <manifest
-    package="drewhamilton.rxpreferences3.dagger" />
+    package="dev.drewhamilton.rxpreferences.dagger" />
