--- dagger/src/main/java/drewhamilton/rxpreferences3/dagger/RxPreferencesModule.kt
@@ -1,14 +1,0 @@
-package drewhamilton.rxpreferences3.dagger
-
-import android.content.SharedPreferences
-import dagger.Module
-import dagger.Provides
-import drewhamilton.rxpreferences3.RxPreferences
-
-@Module
-internal object RxPreferencesModule {
-
-    @JvmStatic
-    @Provides
-    internal fun rxPreferences(sharedPreferences: SharedPreferences) = RxPreferences(sharedPreferences)
-}
