+++ ktx/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
 <manifest
-    package="drewhamilton.rxpreferences3.ktx" />
+    package="dev.drewhamilton.rxpreferences.ktx" />
