+++ ktx/src/test/java/dev/drewhamilton/rxpreferences/PreferenceType.kt
@@ -1,0 +1,10 @@
+package dev.drewhamilton.rxpreferences
+
+enum class PreferenceType(val valueClass: Class<*>) {
+    STRING(String::class.java),
+    STRING_SET(Set::class.java),
+    INT(Int::class.java),
+    LONG(Long::class.java),
+    FLOAT(Float::class.java),
+    BOOLEAN(Boolean::class.java)
+}
