+++ rxpreferences/src/main/java/dev/drewhamilton/rxpreferences/GetPreference.java
@@ -1,0 +1,10 @@
+package dev.drewhamilton.rxpreferences;
+
+import android.content.SharedPreferences;
+import io.reactivex.rxjava3.annotations.NonNull;
+
+interface GetPreference<T> {
+
+    @NonNull
+    T invoke(@NonNull SharedPreferences preferences, @NonNull String key, @NonNull T defaultValue);
+}
