--- rxpreferences/src/main/java/drewhamilton/rxpreferences3/GetPreference.java
@@ -1,10 +1,0 @@
-package drewhamilton.rxpreferences3;
-
-import android.content.SharedPreferences;
-import io.reactivex.rxjava3.annotations.NonNull;
-
-interface GetPreference<T> {
-
-    @NonNull
-    T invoke(@NonNull SharedPreferences preferences, @NonNull String key, @NonNull T defaultValue);
-}
